% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmining.R
\name{tnBooksFromLines}
\alias{tnBooksFromLines}
\title{Title Truenumbers from books}
\usage{
tnBooksFromLines(books)
}
\arguments{
\item{books}{as returned by tidytext}
}
\description{
Title Truenumbers from books
}
