% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.addTag}
\alias{tnum.addTag}
\title{add a tag to a truenumber}
\usage{
tnum.addTag(guid, tag, text = "")
}
\arguments{
\item{guid}{ID of the tn to be tagged}

\item{tag}{path of tag}

\item{text}{reason for application of tag}
}
\description{
add a tag to a truenumber
}
