% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.authorize}
\alias{tnum.authorize}
\title{Connect and authenticate to Truenumbers server}
\usage{
tnum.authorize(ip = "54.166.186.11")
}
\arguments{
\item{ip}{The endpoint address of the server. Default is a public cloud server}

\item{key}{token to authorize API access}
}
\value{
List of numberspaces available on the server. The first one on the list is set as current
}
\description{
Connect and authenticate to Truenumbers server
}
