% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.buildStatement}
\alias{tnum.buildStatement}
\title{Build a truenumber sentence from parts}
\usage{
tnum.buildStatement(
  subject = "something",
  property = "property",
  value = NA,
  error = NA,
  unit = "",
  notes = ""
)
}
\description{
Build a truenumber sentence from parts
}
