% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.buildTnStatement}
\alias{tnum.buildTnStatement}
\title{Build a truenumber sentence from parts}
\usage{
tnum.buildTnStatement(
  subject = "something",
  property = "property",
  value = NA,
  error = NA,
  unit = "",
  notes = ""
)
}
\arguments{
\item{notes}{}
}
\description{
Build a truenumber sentence from parts
}
