% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.createTag}
\alias{tnum.createTag}
\title{Create a tag for application later}
\usage{
tnum.createTag(tag, text = "")
}
\arguments{
\item{tag}{the path or phrase of the tag}

\item{text}{text of the comment associated with tag}
}
\description{
Create a tag for application later
}
