% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.decodeNumericVectorString}
\alias{tnum.decodeNumericVectorString}
\title{Return numeric vector from string "vector(1.23,34.5....)"}
\usage{
tnum.decodeNumericVectorString(nvs)
}
\arguments{
\item{nvs}{}
}
\value{
vector of numbers
}
\description{
Return numeric vector from string "vector(1.23,34.5....)"
}
