% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.deleteTagFromNumberspace}
\alias{tnum.deleteTagFromNumberspace}
\title{Delete a tag AND ALL IT'S APPLICATIONS!}
\usage{
tnum.deleteTagFromNumberspace(tag)
}
\arguments{
\item{tag}{the path of the tag}
}
\description{
Delete a tag AND ALL IT'S APPLICATIONS!
}
