% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.getAttrFromList}
\alias{tnum.getAttrFromList}
\title{Title}
\usage{
tnum.getAttrFromList(obs, attname, defaultValue = NA)
}
\arguments{
\item{obs}{list of tnum objects}

\item{attname}{attribute name}

\item{defaultValue}{default value to use if attribute not present}
}
\value{

}
\description{
Title
}
