% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.graphics.R
\name{tnum.getDatabasePhraseTree}
\alias{tnum.getDatabasePhraseTree}
\title{Get phrase taxonomies}
\usage{
tnum.getDatabasePhraseTree(taxonomy = "subject", pattern = "", levels = 10)
}
\arguments{
\item{taxonomy}{string, one of "subject", "property", or "tags"}

\item{pattern}{a tnum path with path-wildcard #, or string-wildcard * to restrict what tree is returned.}

\item{levels}{integer, how man levels down to extract}
}
\value{
a data.tree Node, suitable for print(returned node) or plot(returned node)
}
\description{
Get phrase taxonomies
}
