% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.graphics.R
\name{tnum.getDBPathList}
\alias{tnum.getDBPathList}
\title{Get phrase taxonomies}
\usage{
tnum.getDBPathList(
  taxonomy = "subject",
  pattern = "",
  levels = NA,
  max = 100,
  start.at = 0
)
}
\arguments{
\item{taxonomy}{string, one of "subject", "property", or "tags"}

\item{pattern}{a tnum path with path-wildcard #, or string-wildcard * to restrict what tree is returned.}

\item{levels}{How deep to go in the taxonomy}

\item{max}{integer, how many results to return max}

\item{start.at}{offset at which to begin returning max results}
}
\value{
a vector of paths in the taxonomy
}
\description{
Get phrase taxonomies
}
