% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.graphics.R
\name{tnum.graphTnumList}
\alias{tnum.graphTnumList}
\title{Make full tnum graph from tnum.query return data frame}
\usage{
tnum.graphTnumList(
  tlist,
  tagpattern = "",
  collectors = list(),
  style = "neato"
)
}
\arguments{
\item{tlist}{list of tnum objects as returned from tnum.query}

\item{tagpattern}{regexp to select tags to include in graph}

\item{collectors}{list of gsub patterns for replacement with ### to aggregate subjects}

\item{style}{what DiagrammeR calls "layout" = tree,nicely,neato,kk, or fr}
}
\description{
Make full tnum graph from tnum.query return data frame
}
