% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.utiities.R
\name{tnum.ingestDataFrameDefault}
\alias{tnum.ingestDataFrameDefault}
\title{default ingest a data frame}
\usage{
tnum.ingestDataFrameDefault(df, root = "", tag = NULL, outfile = "")
}
\arguments{
\item{df}{the data frame, as returned by read.csv() for example}

\item{root}{a string used as the root of the subject for all the TNs}

\item{tag}{a tag to be applied to all the generated TNs}

\item{outfile}{if non-empty, causes TNs and tags to be written to a file, not the server.}
}
\description{
default ingest a data frame
}
\concept{For each row in the dataframe, a TN is generated for each columns, of the form
           <root>dataframe/<row number>:row has <tokenized column name> = <cell value>}
