% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.jsonArray}
\alias{tnum.jsonArray}
\title{make a json array from a list of json objects}
\usage{
tnum.jsonArray(objectList)
}
\arguments{
\item{objectList}{the list of json objects}
}
\value{
json array in R format (jsonlite)
}
\description{
make a json array from a list of json objects
}
