% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.makeNumericVectorString}
\alias{tnum.makeNumericVectorString}
\title{Create a tnum vector value string "vector(23,-34.02...)" from an R vector or list}
\usage{
tnum.makeNumericVectorString(numvec)
}
\arguments{
\item{numvec}{a vector or list}
}
\value{
a string "vector(23,-34.02...)"
}
\description{
Create a tnum vector value string "vector(23,-34.02...)" from an R vector or list
}
