% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.graphics.R
\name{tnum.makePhraseGraphFromPathList}
\alias{tnum.makePhraseGraphFromPathList}
\title{Get a DiagrammeR tree for rendering, from a list of SRD paths}
\usage{
tnum.makePhraseGraphFromPathList(
  pathList = list(),
  rootLabel = "ROOT",
  levels = 10
)
}
\arguments{
\item{levels}{}
}
\value{
Diagrammer graph object
}
\description{
Get a DiagrammeR tree for rendering, from a list of SRD paths
}
