% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.makeTnumJson}
\alias{tnum.makeTnumJson}
\title{Create a JSON truenumber from parts}
\usage{
tnum.makeTnumJson(
  subject = "something",
  property = "property",
  value = NA,
  numeric.error = NA,
  unit = "",
  tags = list(),
  noEmptyStrings = FALSE
)
}
\arguments{
\item{noEmptyStrings}{}
}
\description{
Create a JSON truenumber from parts
}
