% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.makeTnumObject}
\alias{tnum.makeTnumObject}
\title{make a tnum object from numeric values in a tnum data frame}
\usage{
tnum.makeTnumObject(
  subject = "something",
  property = "some-property",
  value,
  error = NA,
  units = NA,
  tags = NA,
  dat = NA,
  gid = NA
)
}
\arguments{
\item{df}{data frame as returned by tnum.query}

\item{numerics}{return df as numeric vector of only numeric values in the df. Defauls is FALSE}
}
\value{
an R object of class tnum
}
\description{
make a tnum object from numeric values in a tnum data frame
}
