% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.maketruenumber}
\alias{tnum.maketruenumber}
\title{Title}
\usage{
tnum.maketruenumber(
  subject = "something",
  property = "property",
  Cvalue = NA,
  Nvalue = NA,
  error = NA,
  units = "",
  tags = list(),
  noEmptyStrings = FALSE
)
}
\arguments{
\item{tags}{}
}
\value{

}
\description{
Title
}
