% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.utiities.R
\name{tnum.parsePhrase}
\alias{tnum.parsePhrase}
\title{parse a phrase as path}
\usage{
tnum.parsePhrase(phr)
}
\arguments{
\item{phr}{phrase string}
}
\value{
path string
}
\description{
parse a phrase as path
}
