% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.utiities.R
\name{tnum.pathLength}
\alias{tnum.pathLength}
\title{Get length of path}
\usage{
tnum.pathLength(path)
}
\arguments{
\item{path}{phrase path}
}
\value{
number of segments
}
\description{
Get length of path
}
