% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.graphics.R
\name{tnum.plotGraph}
\alias{tnum.plotGraph}
\title{Plot DiagrammeR graph}
\usage{
tnum.plotGraph(gph, style = "neato", size = 0)
}
\arguments{
\item{gph}{the graph df}

\item{style}{what DiagrammeR calls "layout" = tree,nicely,neato,kk, or fr}

\item{size}{size of plot in pixels}
}
\value{
result of render_graph() call
}
\description{
Plot DiagrammeR graph
}
