% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postObjects}
\alias{tnum.postObjects}
\title{post a list or vector of tnum objects}
\usage{
tnum.postObjects(objects)
}
\arguments{
\item{objects}{}
}
\value{

}
\description{
post a list or vector of tnum objects
}
