% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postStatement}
\alias{tnum.postStatement}
\title{Post a truenumber statement}
\usage{
tnum.postStatement(
  stmt,
  notes = "",
  tags = list(paste0("source:R:", tnum.dateToken())),
  noreturn = FALSE
)
}
\arguments{
\item{stmt}{well-formed truenumber sentence}

\item{notes}{text description associated with the truenumber}

\item{tags}{a list of tags. Each can be a tag or c("tag","comment"). default is timestamp "R:2022:02:26:12:24:33:EST"}

\item{noreturn}{if TRUE, suppresses return of the created TN}
}
\value{
the new truenumber JSON object (or NULL if suppressed)
}
\description{
Post a truenumber statement
}
