% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postTnums}
\alias{tnum.postTnums}
\title{post a list or vector of tnum objects}
\usage{
tnum.postTnums(objects)
}
\arguments{
\item{objects}{}
}
\value{

}
\description{
post a list or vector of tnum objects
}
