% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postTruenumber}
\alias{tnum.postTruenumber}
\title{Post a single truenumber from parts}
\usage{
tnum.postTruenumber(
  subject = "something",
  property = "property",
  string.value = NA,
  numeric.value = NA,
  numeric.error = NA,
  units = "",
  tags = list(),
  noEmptyStrings = FALSE
)
}
\arguments{
\item{tags}{}
}
\value{

}
\description{
Post a single truenumber from parts
}
