% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postTruenumberObjects}
\alias{tnum.postTruenumberObjects}
\title{post a list or vector of tnum objects}
\usage{
tnum.postTruenumberObjects(objects, noEmptyStrings = FALSE)
}
\arguments{
\item{noEmptyStrings}{}
}
\value{

}
\description{
post a list or vector of tnum objects
}
