% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postTruenumbers}
\alias{tnum.postTruenumbers}
\title{Create many truenumbers from lists of parts}
\usage{
tnum.postTruenumbers(
  subject,
  property,
  string.value,
  numeric.value,
  numeric.error,
  units,
  tags,
  noEmptyStrings = FALSE
)
}
\arguments{
\item{noEmptyStrings}{if true doesn't post TNs with empty values}
}
\value{

}
\description{
Create many truenumbers from lists of parts
}
