% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.queryResultToDataframe}
\alias{tnum.queryResultToDataframe}
\title{Convert tnum query result to a data frame}
\usage{
tnum.queryResultToDataframe(result, max)
}
\arguments{
\item{result}{from a tnum API query}

\item{max}{maximum rows to return}
}
\value{

}
\description{
Convert tnum query result to a data frame
}
