% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.queryResultToObjects}
\alias{tnum.queryResultToObjects}
\title{Convert tnum query result to an object list}
\usage{
tnum.queryResultToObjects(result, SI = FALSE, max = 100)
}
\arguments{
\item{result}{from a tnum API query}

\item{SI}{if TRUE, the object will reflect numerical SI values}

\item{max}{maximum rows to return}
}
\value{
list of tnum objects
}
\description{
Convert tnum query result to an object list
}
