% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.removeTag}
\alias{tnum.removeTag}
\title{remove a tag from a truenumber}
\usage{
tnum.removeTag(tag, number)
}
\arguments{
\item{tag}{matching the tag(s) to remove}

\item{number}{GUID of the number}
}
\description{
remove a tag from a truenumber
}
