% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.setspace}
\alias{tnum.setspace}
\title{Set a particular numberspace as current}
\usage{
tnum.setspace(name = "testspace")
}
\arguments{
\item{name}{Name of numberspace}
}
\value{
-none-
}
\description{
Set a particular numberspace as current
}
