% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.utiities.R
\name{tnum.subPath}
\alias{tnum.subPath}
\title{Get subpath of length N}
\usage{
tnum.subPath(path, n = 1)
}
\arguments{
\item{path}{the path}

\item{n}{length of desired sub-path}
}
\value{
first n terms of path.
}
\description{
Get subpath of length N
}
