% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.tagByGuids}
\alias{tnum.tagByGuids}
\title{Add a column of single tags element-wise to list of tnums by GUID}
\usage{
tnum.tagByGuids(gids = c(), adds = c())
}
\arguments{
\item{gids}{guids of tnums to tag}

\item{adds}{tags to add}
}
\value{
result of API call
}
\description{
Add a column of single tags element-wise to list of tnums by GUID
}
