% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.tnumObjectsToDf}
\alias{tnum.tnumObjectsToDf}
\title{make data frame from list of tnum objects}
\usage{
tnum.tnumObjectsToDf(objs)
}
\arguments{
\item{objs}{tnum list}
}
\value{
data frame
}
\description{
make data frame from list of tnum objects
}
