% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.writeTnFile}
\alias{tnum.writeTnFile}
\title{Post a TN file to the server}
\usage{
tnum.writeTnFile(infile)
}
\arguments{
\item{infile}{path to the file, usually the same given to the ingest function}
}
\value{
post result
}
\description{
Post a TN file to the server
}
