% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.buildStatement}
\alias{tnum.buildStatement}
\title{Build a truenumber statement from parts}
\usage{
tnum.buildStatement(
  subject = "something",
  property = "property",
  value = NA,
  error = NA,
  unit = ""
)
}
\arguments{
\item{subject}{TN subject}

\item{property}{TN property}

\item{value}{TN value}

\item{error}{std error for numerical value}

\item{unit}{units of measure like m/s^2}
}
\description{
Build a truenumber statement from parts
}
