% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.utiities.R
\name{tnum.ingestDataFrame}
\alias{tnum.ingestDataFrame}
\title{ingest a data frame as verbatim truenumbers}
\usage{
tnum.ingestDataFrame(
  df,
  subjectRoot,
  idType = "/row:",
  idColumn,
  tag = "origin:R/tnum"
)
}
\arguments{
\item{df}{the data frame, as returned by read.csv() for example}

\item{subjectRoot}{string used as root of subject path. It should not end with : or /}

\item{idType}{string appended to subjectRoot to prefix the idColumn token. Must start and end with : or /}

\item{idColumn}{name of a column to be used at the end of the subject path to identify the row}

\item{tag}{path used to tag all the created truenumbers}
}
\description{
ingest a data frame as verbatim truenumbers
}
