% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.utiities.R
\name{tnum.ingestDataFrame}
\alias{tnum.ingestDataFrame}
\title{ingest a data frame as  truenumbers from templates}
\usage{
tnum.ingestDataFrame(df, templates = list(), outfile = "")
}
\arguments{
\item{df}{the data frame, as returned by read.csv() for example}

\item{templates}{a list of string pairs, the first is a TN template, the second is a tag list template
TN template is a truenumber sentence including "macros" to be replace by row data values
Tag list template is a comma-separated list of tag paths, inscluding macros as well
Macros are of the form $funcName(column name), or $(column name).  The column value is substituted
for the macro.  If a function name is present, that function processes the column value before substitution.}

\item{outfile}{if non-empty, causes TNs and tags to be written to a file, not the server.}
}
\description{
ingest a data frame as  truenumbers from templates
}
\concept{For each row in the dataframe, a set of tagged truenumbers can be generated and posted based ona list of templates.
         Templates are TNs and tags specified as strings, containing special macros that refer to column values of the row.
         Thus, for example, ingesting a data frame of 10 columns and 100 rows, using a list of 3 templates, would produce
         3 TNs per row, or 300 TNs.}
