% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.graphics.R
\name{tnum.makeTnumPhraseGraph}
\alias{tnum.makeTnumPhraseGraph}
\title{Make full tnum graph from tnum.query return data frame}
\usage{
tnum.makeTnumPhraseGraph(tlist, tagpattern = "", collectors = list())
}
\arguments{
\item{tlist}{list of tnum objects as returned from tnum.query}

\item{tagpattern}{regexp to select tags to include in graph}

\item{collectors}{list of gsub patterns for replacement with ### to aggregate subjects}
}
\value{
returns a DiagrammeR graph
}
\description{
Make full tnum graph from tnum.query return data frame
}
