% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.postStatement}
\alias{tnum.postStatement}
\title{post a list or vector of tnum objects}
\usage{
tnum.postStatement(sentence, tags = list())
}
\arguments{
\item{tags}{}
}
\description{
post a list or vector of tnum objects
}
