% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.query}
\alias{tnum.query}
\title{Query the truenumber server}
\usage{
tnum.query(query = "* has *", max = 10, start = 0, fast = FALSE)
}
\arguments{
\item{query}{string in tnum query language}

\item{max}{maximum number of truenumbers to return}

\item{start}{begin return with this sequence number}
}
\value{
list of json truenumber as received from server
}
\description{
Query the truenumber server
}
