% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.query}
\alias{tnum.query}
\title{Query the truenumber DB}
\usage{
tnum.query(query = "* has *", SI = FALSE, max = 10, start = 0)
}
\arguments{
\item{query}{string in tnum query language}

\item{SI}{if TRUE, returns the SI value}

\item{max}{maximum nuber of truenumbers to return}

\item{start}{begin return with this sequence number}
}
\value{
a list of tnum class vaues
}
\description{
Query the truenumber DB
}
