% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnum.R
\name{tnum.tagByGuids}
\alias{tnum.tagByGuids}
\title{Add and remove tags to tnums specified by a list of their GUID's}
\usage{
tnum.tagByGuids(gids = c(), adds = c(), removes = c())
}
\arguments{
\item{gids}{guids of tnums to tag}

\item{adds}{tags to add}

\item{removes}{tags to remove}
}
\value{
result of API call
}
\description{
Add and remove tags to tnums specified by a list of their GUID's
}
